ContextAlgo : Remove in favour of ScriptNodeAlgo
We deprecated ScriptNodeAlgo in 1.4.13.0, and the alternative ScriptNodeAlgo API has been available since then as well.

Here we just move the implementation from ContextAlgo.cpp to ScriptNodeAlgo.cpp, and add a bit of missing test coverage. Next we will change the implementation so that is doesn't abuse contexts for UI state storage.
