+++ python/GafferSceneUITest/__init__.py
@@ -46,7 +46,6 @@ from .ShaderUITest import ShaderUITest
 from .TranslateToolTest import TranslateToolTest
 from .ScaleToolTest import ScaleToolTest
 from .RotateToolTest import RotateToolTest
-from .ContextAlgoTest import ContextAlgoTest
 from .CameraToolTest import CameraToolTest
 from .VisualiserTest import VisualiserTest
 from .TransformToolTest import TransformToolTest
